Micromachines

Documentation

constroi :: Caminho -> Mapa

Dado um caminho qualquer, constroi o mapa associado

caminho2mapa :: Posicao -> Posicao -> Orientacao -> Altura -> Caminho -> Tabuleiro

validaMapa :: Mapa -> Bool

Verifica se um dado mapa é válido:

começando da partida volta-se a chegar à partida

não há peças de caminho não utilizadas

peças só podem estar ligadas por lados com a mesma altura

posicoesDoCaminho :: Mapa -> Maybe [Posicao]

restoComLava :: Tabuleiro -> Maybe [Posicao] -> Bool

voltaEmLava :: Tabuleiro -> Bool

colide :: Tabuleiro -> Tempo -> Carro -> Maybe Carro

Dado um ponto e uma velocidade inciais, e uma duração de tempo, se houver uma colisão ou uma queda, calcula um ponto de chegada e a nova velocidade assumindo que:

a velocidade se mantem constante;

se bater numa parede a velocidade é reflectida e a sua norma é mantida

se cair ou passar pela lava, a velocidade final é nula e a posição final é o centro da peça onde caíu ou atravessou para a lava.

caso contrário (não há colisões nem quedas) devolve Nothing

low :: Double -> Double

high :: Double -> Double

parede :: Peca -> Ponto -> Altura -> Bool

transita :: Peca -> Ponto -> Ponto -> Ponto

Recebe uma peca, um ponto inicial e um ponto final. Calcula a primeira intersecção com uma linha.

Assume que está numa posição válida, ou seja, numa posição de estrada.

transitaAresta :: Ponto -> Ponto -> Ponto

Devolve a próxima aresta no caminho

colideDentro :: Peca -> Ponto -> Ponto -> Maybe Ponto

interNaPeca :: Peca -> Posicao -> Maybe Linha

p :: (Double, Double)

v :: (Double, Double)

t :: Tempo

f :: Ponto

type Linha = (Ponto, Ponto)

intersecta :: Linha -> Linha -> Maybe Ponto

Calcula o ponto de intercepção entre duas linhas

intersectaL :: Ponto -> Ponto -> Ponto -> Ponto -> Maybe Ponto

closestPontoOnL :: Ponto -> Ponto -> Ponto -> Double

ponto2pos :: Ponto -> Posicao

pos2ponto :: Posicao -> Ponto

final :: Ponto -> Velocidade -> Tempo -> Ponto

Dado um ponto, um vector velocidade e um intervalo de tempo, devolve o ponto final

(/+) :: Ponto -> Ponto -> Ponto

(/-) :: Ponto -> Ponto -> Ponto

(/*) :: Ponto -> Ponto -> Ponto

(//) :: Ponto -> Ponto -> Ponto

linhaLava :: Int -> [Peca]

mapaLava :: (Int, Int) -> Tabuleiro

getPeca :: Posicao -> Tabuleiro -> Peca

setPeca :: Posicao -> Peca -> Tabuleiro -> Tabuleiro

isLava :: Peca -> Bool

isCurva :: Peca -> Bool

altura :: Peca -> Altura

inBounds :: Posicao -> Posicao -> Bool

inTabuleiro :: Posicao -> Tabuleiro -> Bool

linhasMesmoTamanho :: Tabuleiro -> Bool

minMax :: Posicao -> (Posicao, Posicao) -> (Posicao, Posicao)

pmax :: Posicao -> Posicao -> Posicao

pmin :: Posicao -> Posicao -> Posicao

rodaDir :: Orientacao -> Orientacao

rodaEsq :: Orientacao -> Orientacao

inverte :: Orientacao -> Orientacao

proxPos :: Posicao -> Orientacao -> Posicao

pecaCompativel :: Orientacao -> Peca -> Bool

acessoOk :: Orientacao -> Peca -> Altura -> Bool

proxAlt :: Peca -> Orientacao -> Altura

saidaCurva :: Orientacao -> Orientacao -> Orientacao

proxByPeca :: Peca -> Orientacao -> Posicao -> (Posicao, Orientacao, Altura)

sizeMapa :: Mapa -> Posicao

filtraPeca :: Tabuleiro -> (Peca -> Bool) -> [Posicao]