Tarefa4_2017li1g183

Description

Módulo para a realização da Tarefa 4 de LI1 em 2017/18.

Documentation

testesT4 :: [(Tempo, Jogo, Acao)]

O testes a serem considerados pelo sistema de feedback para a função atualiza.

atualiza

Arguments

:: Tempo
a duração da ação

-> Jogo
o estado do jogo

-> Int
o identificador do jogador

-> Acao
a ação tomada pelo jogador

-> Jogo
o estado atualizado do jogo

Função usada para atualizar o estado do jogo dadas as ações de um jogador num determinado período de tempo.